requirements.txt
yt-dlp
openai-whisper
google-genai
ffmpeg-python
